pyproject.toml
[project]
name = "kermi2mqtt"
version = "0.0.1"
description = "Modbus-to-MQTT bridge for Kermi heat pumps with Home Assistant auto-discovery"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "Apache-2.0" }
authors = [
    { name = "kermi2mqtt contributors" }
]
keywords = ["kermi", "heat-pump", "modbus", "mqtt", "home-assistant", "iot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Home Automation",
    "Topic :: System :: Hardware :: Hardware Drivers",
]

dependencies = [
    "kermi-xcenter>=0.2.2",
    "aiomqtt>=2.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "black>=24.0.0",
]

[project.scripts]
kermi2mqtt = "kermi2mqtt.__main__:main"

[project.urls]
Homepage = "https://github.com/yourusername/kermi2mqtt"
Repository = "https://github.com/yourusername/kermi2mqtt"
Issues = "https://github.com/yourusername/kermi2mqtt/issues"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["kermi2mqtt", "kermi2mqtt.models"]
package-dir = {"" = "src"}

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py312"
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "ANN", # flake8-annotations
    "ASYNC", # flake8-async
    "S",   # flake8-bandit
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "DTZ", # flake8-datetimez
    "T10", # flake8-debugger
    "EM",  # flake8-errmsg
    "ISC", # flake8-implicit-str-concat
    "ICN", # flake8-import-conventions
    "PIE", # flake8-pie
    "PT",  # flake8-pytest-style
    "Q",   # flake8-quotes
    "RSE", # flake8-raise
    "RET", # flake8-return
    "SIM", # flake8-simplify
    "TID", # flake8-tidy-imports
    "ARG", # flake8-unused-arguments
    "PTH", # flake8-use-pathlib
    "ERA", # eradicate
    "PL",  # pylint
    "RUF", # ruff-specific rules
]
ignore = [
    "ANN101", # Missing type annotation for self
    "ANN102", # Missing type annotation for cls
    "S101",   # Use of assert (needed for pytest)
]

[tool.ruff.per-file-ignores]
"tests/**/*.py" = ["S101", "PLR2004", "ANN"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "kermi_xcenter.*",
    "aiomqtt.*",
]
ignore_missing_imports = true

[tool.black]
line-length = 100
target-version = ["py312"]
include = '\.pyi?$'
